Fix crawler for dhlottery's rebuilt site, add RestClient bean, fix 401 bug
The site behind the design doc's common.do?method=getLottoNumber API was
rebuilt as an SPA; that legacy endpoint now 404s. Found the real replacement
by inspecting network traffic in a browser:

  GET /lt645/selectPstLt645InfoNew.do?srchDir=center&srchLtEpsd={drawNo}

It returns a JSON envelope with a `data.list` array of up to 10 draws
around the requested one (field names, date format, and "not found" vs.
"fail" semantics all changed too). Rewrote the DTOs/parser/client and
their tests against this shape; verified end-to-end against a live
Supabase instance (full 1233-draw backfill, repeat calls no-op safely).

Also:
- RestClient.Builder isn't autoconfigured by spring-boot-starter-webmvc
  in this Spring Boot 4.1.0 setup, so DhLotteryClient's constructor
  dependency failed to resolve; added an explicit @Bean for it.
- lotto_draws numeric columns changed from smallint to integer -
  Hibernate schema validation rejected the smallint/Integer mismatch.
- /api/crawl returned 400 instead of 401 when the Authorization header
  was missing (required=true on @RequestHeader throws before the auth
  check runs); made the header optional so the check actually executes.

+++ db/migrations/0001_create_lotto_draws.sql
@@ -1,12 +1,12 @@
 create table if not exists lotto_draws (
   draw_no integer primary key,
   draw_date date not null,
-  num1 smallint not null,
-  num2 smallint not null,
-  num3 smallint not null,
-  num4 smallint not null,
-  num5 smallint not null,
-  num6 smallint not null,
-  bonus_num smallint not null,
+  num1 integer not null,
+  num2 integer not null,
+  num3 integer not null,
+  num4 integer not null,
+  num5 integer not null,
+  num6 integer not null,
+  bonus_num integer not null,
   created_at timestamptz not null default now()
 );
